requirements.txt
packageurl_python==0.8.7
requests==2.22.0
xmlschema==1.0.16
setuptools==18.5
requirements_parser==0.2.0
packaging==19.2
chardet==3.0.4
pytest==4.6.9